src/requirements.txt
google-genai>=1.0.0,<2.0.0
python-escpos>=3.1.0,<4.0.0
jira>=3.10.5,<4.0.0
doppler-sdk>=1.3.0,<2.0.0
pyusb>=1.3.1,<2.0.0
